Remove unnecessary catch clauses for SAXNotRecognizedException and SAXNotSupportedException
Removed the catch clauses for org.xml.sax.SAXNotRecognizedException and org.xml.sax.SAXNotSupportedException from the try statement in JavaProvider.java, as these exceptions are never thrown in the body of the try block. Also removed the unused import statements for these exceptions. This resolves the specified bug and cleans up the exception handling.